Handle H2D scales in bf16 datatype (#1474)

+++ tests/pytest_working/any_mode/test_hpu_fp8_ops_any_mode.py
@@ -1175,9 +1175,9 @@ def test_h2d_scales(src_dtype, batched_tensors, fuse_cast):
             b /= 4.0
         ah = a.to("hpu")
         bh = b.to("hpu")
-        sa = torch.tensor(sa_val)
-        sb = torch.tensor(sb_val)
-        so = torch.tensor(scale_out)
+        sa = torch.tensor(sa_val, dtype=src_dtype)
+        sb = torch.tensor(sb_val, dtype=src_dtype)
+        so = torch.tensor(scale_out, dtype=src_dtype)
 
         return a, b, ah, bh, sa, sb, so
